Adds a concept of user context and user owned data. I don't really like the provider being responsible for accessing user context, will likely change this.

+++ Binner/Binner.Common/StorageProviders/SqlServerStorageProvider.cs
@@ -1,16 +1,12 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Data;
-using System.Linq;
-using System.Reflection;
-using System.Text;
-using System.Threading.Tasks;
-using Binner.Common.Extensions;
+﻿using Binner.Common.Extensions;
 using Binner.Common.Models;
 using Microsoft.Data.SqlClient;
-using TypeSupport;
+using System;
+using System.Collections.Generic;
+using System.Data;
+using System.Data.SqlTypes;
+using System.Linq;
+using System.Threading.Tasks;
 using TypeSupport.Extensions;
 
 namespace Binner.Common.StorageProviders
@@ -23,16 +19,19 @@ namespace Binner.Common.StorageProviders
         public const string ProviderName = "SqlServer";
 
         private readonly SqlServerStorageConfiguration _config;
+        private readonly RequestContextAccessor _requestContext;
         private bool _isDisposed;
 
-        public SqlServerStorageProvider(IDictionary<string, string> config)
+        public SqlServerStorageProvider(IDictionary<string, string> config, RequestContextAccessor requestContext)
         {
             _config = new SqlServerStorageConfiguration(config);
+            _requestContext = requestContext;
             Task.Run(async () => await GenerateDatabaseIfNotExistsAsync<IBinnerDb>()).GetAwaiter().GetResult();
         }
 
         public async Task<Part> AddPartAsync(Part part)
         {
+            part.UserId = GetUserContext(u => u.UserId);
             var query =
 $@"INSERT INTO Parts (Quantity, LowStockThreshold, PartNumber, DigiKeyPartNumber, MouserPartNumber, Description, PartTypeId, ProjectId, Keywords, DatasheetUrl, Location, BinNumber, BinNumber2, UserId) 
 output INSERTED.PartId 
@@ -43,43 +42,47 @@ VALUES(@Quantity, @LowStockThreshold, @PartNumber, @DigiKeyPartNumber, @MouserPa
 
         public async Task<Project> AddProjectAsync(Project project)
         {
+            project.UserId = GetUserContext(u => u.UserId);
             var query =
-            $@"INSERT INTO Projects (Name, Description, DateCreatedUtc, UserId) 
+            $@"INSERT INTO Projects (Name, Description, Location, DateCreatedUtc, UserId) 
 output INSERTED.ProjectId 
-VALUES(@Name, @Description, @DateCreatedUtc, @UserId);
+VALUES(@Name, @Description, @Location, @DateCreatedUtc, @UserId);
 ";
             return await InsertAsync<Project, long>(query, project, (x, key) => { x.ProjectId = key; });
         }
 
         public async Task<bool> DeletePartAsync(Part part)
         {
-            var query = $"DELETE FROM Projects WHERE PartId = @PartId AND UserId = @UserId;";
+            part.UserId = GetUserContext(u => u.UserId);
+            var query = $"DELETE FROM Parts WHERE PartId = @PartId AND (@UserId IS NULL OR UserId = @UserId);";
             return await ExecuteAsync<Part>(query, part) > 0;
         }
 
         public async Task<bool> DeleteProjectAsync(Project project)
         {
-            var query = $"DELETE FROM Projects WHERE ProjectId = @ProjectId AND UserId = @UserId;";
+            project.UserId = GetUserContext(u => u.UserId);
+            var query = $"DELETE FROM Projects WHERE ProjectId = @ProjectId AND (@UserId IS NULL OR UserId = @UserId);";
             return await ExecuteAsync<Project>(query, project) > 0;
         }
 
         public async Task<ICollection<SearchResult<Part>>> FindPartsAsync(string keywords)
         {
-            var query = $"SELECT * FROM Parts WHERE UserId = @UserId AND PartNumber LIKE @Keywords OR DigiKeyPartNumber LIKE @Keywords OR Description LIKE @Keywords OR Keywords LIKE @Keywords OR Location LIKE @Keywords OR BinNumber LIKE @Keywords OR BinNumber2 LIKE @Keywords;";
-            var result = await SqlQueryAsync<Part>(query, new { Keywords = keywords });
+            var query = $"SELECT * FROM Parts WHERE (@UserId IS NULL OR UserId = @UserId) AND PartNumber LIKE @Keywords OR DigiKeyPartNumber LIKE @Keywords OR Description LIKE @Keywords OR Keywords LIKE @Keywords OR Location LIKE @Keywords OR BinNumber LIKE @Keywords OR BinNumber2 LIKE @Keywords;";
+            var result = await SqlQueryAsync<Part>(query, new { Keywords = keywords, UserId = GetUserContext(u => u.UserId) });
             return result.Select(x => new SearchResult<Part>(x, 100)).ToList();
         }
 
         public async Task<OAuthCredential> GetOAuthCredentialAsync(string providerName)
         {
-            var query = $"SELECT * FROM OAuthCredentials WHERE UserId = @UserId AND Provider = @ProviderName;";
-            var result = await SqlQueryAsync<OAuthCredential>(query, new { ProviderName = providerName });
+            var query = $"SELECT * FROM OAuthCredentials WHERE Provider = @ProviderName AND (@UserId IS NULL OR UserId = @UserId);";
+            var result = await SqlQueryAsync<OAuthCredential>(query, new { ProviderName = providerName, UserId = GetUserContext(u => u.UserId) });
             return result.FirstOrDefault();
         }
 
         public async Task<PartType> GetOrCreatePartTypeAsync(PartType partType)
         {
-            var query = $"SELECT PartTypeId FROM PartTypes WHERE Name = @Name;";
+            partType.UserId = GetUserContext(u => u.UserId);
+            var query = $"SELECT PartTypeId FROM PartTypes WHERE Name = @Name AND (@UserId IS NULL OR UserId = @UserId);";
             var result = await SqlQueryAsync<PartType>(query, partType);
             if (result.Any())
             {
@@ -98,15 +101,15 @@ VALUES (@ParentPartTypeId, @Name, @UserId);";
 
         public async Task<Part> GetPartAsync(long partId)
         {
-            var query = $"SELECT * FROM Parts WHERE PartId = @PartId;";
-            var result = await SqlQueryAsync<Part>(query, new { PartId = partId });
+            var query = $"SELECT * FROM Parts WHERE PartId = @PartId AND (@UserId IS NULL OR UserId = @UserId);";
+            var result = await SqlQueryAsync<Part>(query, new { PartId = partId, UserId = GetUserContext(u => u.UserId) });
             return result.FirstOrDefault();
         }
 
         public async Task<Part> GetPartAsync(string partNumber)
         {
-            var query = $"SELECT * FROM Parts WHERE PartNumber = @PartNumber;";
-            var result = await SqlQueryAsync<Part>(query, new { PartNumber = partNumber });
+            var query = $"SELECT * FROM Parts WHERE PartNumber = @PartNumber AND (@UserId IS NULL OR UserId = @UserId);";
+            var result = await SqlQueryAsync<Part>(query, new { PartNumber = partNumber, UserId = GetUserContext(u => u.UserId) });
             return result.FirstOrDefault();
         }
 
@@ -119,45 +122,46 @@ VALUES (@ParentPartTypeId, @Name, @UserId);";
 
         public async Task<Project> GetProjectAsync(long projectId)
         {
-            var query = $"SELECT * FROM Projects WHERE ProjectId = @ProjectId;";
-            var result = await SqlQueryAsync<Project>(query, new { ProjectId = projectId });
+            var query = $"SELECT * FROM Projects WHERE ProjectId = @ProjectId AND (@UserId IS NULL OR UserId = @UserId);";
+            var result = await SqlQueryAsync<Project>(query, new { ProjectId = projectId, UserId = GetUserContext(u => u.UserId) });
             return result.FirstOrDefault();
         }
 
         public async Task<Project> GetProjectAsync(string projectName)
         {
-            var query = $"SELECT * FROM Projects WHERE Name = @Name;";
-            var result = await SqlQueryAsync<Project>(query, new { Name = projectName });
+            var query = $"SELECT * FROM Projects WHERE Name = @Name AND (@UserId IS NULL OR UserId = @UserId);";
+            var result = await SqlQueryAsync<Project>(query, new { Name = projectName, UserId = GetUserContext(u => u.UserId) });
             return result.FirstOrDefault();
         }
 
         public async Task<ICollection<Project>> GetProjectsAsync()
         {
-            var query = $"SELECT * FROM Projects;";
+            var query = $"SELECT * FROM Projects WHERE (@UserId IS NULL OR UserId = @UserId);";
             var result = await SqlQueryAsync<Project>(query);
             return result.ToList();
         }
 
         public async Task RemoveOAuthCredentialAsync(string providerName)
         {
-            var query = $"DELETE FROM OAuthCredentials WHERE Provider = @Provider;";
-            await ExecuteAsync<object>(query, new { Provider = providerName });
+            var query = $"DELETE FROM OAuthCredentials WHERE Provider = @Provider AND (@UserId IS NULL OR UserId = @UserId);";
+            await ExecuteAsync<object>(query, new { Provider = providerName, UserId = GetUserContext(u => u.UserId) });
         }
 
         public async Task<OAuthCredential> SaveOAuthCredentialAsync(OAuthCredential credential)
         {
-            var query = @"SELECT Provider FROM OAuthCredentials WHERE Provider = @Provider;";
+            credential.UserId = GetUserContext(u => u.UserId);
+            var query = @"SELECT Provider FROM OAuthCredentials WHERE Provider = @Provider AND (@UserId IS NULL OR UserId = @UserId);";
             var result = await SqlQueryAsync<OAuthCredential>(query, credential);
             if (result.Any())
             {
-                query = $@"UPDATE OAuthCredentials SET AccessToken = @AccessToken, RefreshToken = @RefreshToken, DateCreatedUtc = @DateCreatedUtc, DateExpiresUtc = @DateExpiresUtc WHERE Provider = @Provider;";
+                query = $@"UPDATE OAuthCredentials SET AccessToken = @AccessToken, RefreshToken = @RefreshToken, DateCreatedUtc = @DateCreatedUtc, DateExpiresUtc = @DateExpiresUtc WHERE Provider = @Provider AND (@UserId IS NULL OR UserId = @UserId);";
                 await ExecuteAsync<object>(query, credential);
             }
             else
             {
                 query =
-$@"INSERT INTO OAuthCredentials (Provider, AccessToken, RefreshToken, DateCreatedUtc, DateExpiresUtc) 
-VALUES (Provider, AccessToken, RefreshToken, DateCreatedUtc, DateExpiresUtc);";
+$@"INSERT INTO OAuthCredentials (Provider, AccessToken, RefreshToken, DateCreatedUtc, DateExpiresUtc, UserId) 
+VALUES (@Provider, @AccessToken, @RefreshToken, @DateCreatedUtc, @DateExpiresUtc, @UserId);";
                 await InsertAsync<OAuthCredential, string>(query, credential, (x, key) => { });
             }
             return credential;
@@ -165,11 +169,12 @@ VALUES (Provider, AccessToken, RefreshToken, DateCreatedUtc, DateExpiresUtc);";
 
         public async Task<Part> UpdatePartAsync(Part part)
         {
-            var query = $"SELECT PartId FROM Parts WHERE PartId = @PartId AND UserId = @UserId;";
+            part.UserId = GetUserContext(u => u.UserId);
+            var query = $"SELECT PartId FROM Parts WHERE PartId = @PartId AND (@UserId IS NULL OR UserId = @UserId);";
             var result = await SqlQueryAsync<Part>(query, part);
             if (result.Any())
             {
-                query = $"UPDATE Parts SET Quantity = @Quantity, LowStockThreshold = @LowStockThreshold, PartNumber = @PartNumber, DigiKeyPartNumber = @DigiKeyPartNumber, MouserPartNumber = @MouserPartNumber, Description = @Description, PartTypeId = @PartTypeId, ProjectId = @ProjectId, Keywords = @Keywords, DatasheetUrl = @DatasheetUrl, Location = @Location, BinNumber = @BinNumber, BinNumber2 = @BinNumber2 WHERE PartId = @PartId AND UserId = @UserId;";
+                query = $"UPDATE Parts SET Quantity = @Quantity, LowStockThreshold = @LowStockThreshold, PartNumber = @PartNumber, DigiKeyPartNumber = @DigiKeyPartNumber, MouserPartNumber = @MouserPartNumber, Description = @Description, PartTypeId = @PartTypeId, ProjectId = @ProjectId, Keywords = @Keywords, DatasheetUrl = @DatasheetUrl, Location = @Location, BinNumber = @BinNumber, BinNumber2 = @BinNumber2 WHERE PartId = @PartId AND (@UserId IS NULL OR UserId = @UserId);";
                 await ExecuteAsync<Part>(query, part);
             }
             else
@@ -181,11 +186,12 @@ VALUES (Provider, AccessToken, RefreshToken, DateCreatedUtc, DateExpiresUtc);";
 
         public async Task<Project> UpdateProjectAsync(Project project)
         {
-            var query = $"SELECT ProjectId FROM Projects WHERE ProjectId = @ProjectId AND UserId = @UserId;";
+            project.UserId = GetUserContext(u => u.UserId);
+            var query = $"SELECT ProjectId FROM Projects WHERE ProjectId = @ProjectId AND (@UserId IS NULL OR UserId = @UserId);";
             var result = await SqlQueryAsync<Project>(query, project);
             if (result.Any())
             {
-                query = $"UPDATE Projects SET Name = @Name, Description = @Description, DateCreatedUtc = @DateCreatedUtc WHERE ProjectId = @ProjectId AND UserId = @UserId;";
+                query = $"UPDATE Projects SET Name = @Name, Description = @Description, Location = @Location WHERE ProjectId = @ProjectId AND (@UserId IS NULL OR UserId = @UserId);";
                 await ExecuteAsync<Project>(query, project);
             }
             else
@@ -193,6 +199,24 @@ VALUES (Provider, AccessToken, RefreshToken, DateCreatedUtc, DateExpiresUtc);";
                 throw new ArgumentException($"Record not found for {nameof(Project)} = {project.ProjectId}");
             }
             return project;
+        }
+
+        private Nullable<T> GetUserContext<T>(Func<UserContext, T> userSelector)
+            where T : struct
+        {
+            var userContext = _requestContext.GetUserContext();
+            if (userContext != null)
+                return userSelector.Invoke(userContext);
+            return new Nullable<T>(default(T));
+        }
+
+        private T GetUserContext<T>(Func<UserContext, T> userSelector, T defaultValue)
+            where T : class
+        {
+            var userContext = _requestContext.GetUserContext();
+            if (userContext != null)
+                return userSelector.Invoke(userContext);
+            return default(T) ?? defaultValue;
         }
 
         private async Task<T> InsertAsync<T, TKey>(string query, T parameters, Action<T, TKey> keySetter)
@@ -232,7 +256,10 @@ VALUES (Provider, AccessToken, RefreshToken, DateCreatedUtc, DateExpiresUtc);";
                         foreach (var prop in type.Properties)
                         {
                             if (reader.HasColumn(prop.Name))
-                                newObj.SetPropertyValue(prop.PropertyInfo, reader[prop.Name]);
+                            {
+                                var val = MapToPropertyValue(reader[prop.Name], prop.Type);
+                                newObj.SetPropertyValue(prop.PropertyInfo, val);
+                            }
                         }
                         results.Add(newObj);
                     }
@@ -266,13 +293,28 @@ VALUES (Provider, AccessToken, RefreshToken, DateCreatedUtc, DateExpiresUtc);";
             foreach (var property in properties)
             {
                 var propertyValue = record.GetPropertyValue(property);
-                var propertyMapped = MapPropertyValue(propertyValue);
+                var propertyMapped = MapFromPropertyValue(propertyValue);
                 parameters.Add(new SqlParameter(property.Name, propertyMapped));
             }
             return parameters.ToArray();
         }
 
-        private object MapPropertyValue(object obj)
+        private object MapToPropertyValue(object obj, Type destinationType)
+        {
+            if (obj == DBNull.Value) return null;
+
+            var objType = destinationType.GetExtendedType();
+            switch (objType)
+            {
+                case var p when p.IsCollection:
+                case var a when a.IsArray:
+                    return obj.ToString().Split(new string[] { "," }, StringSplitOptions.RemoveEmptyEntries);
+                default:
+                    return obj;
+            }
+        }
+
+        private object MapFromPropertyValue(object obj)
         {
             if (obj == null) return DBNull.Value;
 
@@ -282,6 +324,10 @@ VALUES (Provider, AccessToken, RefreshToken, DateCreatedUtc, DateExpiresUtc);";
                 case var p when p.IsCollection:
                 case var a when a.IsArray:
                     return string.Join(",", ((ICollection<object>)obj).Select(x => x.ToString()).ToArray());
+                case var p when p.Type == typeof(DateTime):
+                    if (((DateTime)obj) == DateTime.MinValue)
+                        return SqlDateTime.MinValue.Value;
+                    return obj;
                 default:
                     return obj;
             }
